requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
PyJWT
pydantic[email]
python-decouple